repos conectados con vistas para hacer las estadisticas

+++ Registro Viajes/RepositorioViaje.cs
@@ -33,7 +33,7 @@ namespace UberFrba.Registro_Viajes
         #region builder del objeto viaje
         public override Viaje BuilderEntityFromDataRow(DataRow dr)
         {
-            Viaje viaje = new Viaje(Convert.ToInt16(dr[0]), Convert.ToInt16(dr[1]), dr[2].ToString(), Convert.ToInt16(dr[3]), Convert.ToDouble(dr[4]), Convert.ToDateTime(dr[5]), Convert.ToDateTime(dr[6]), Convert.ToInt16(dr[7]), Convert.ToInt16(dr[8]));
+            Viaje viaje = new Viaje(Convert.ToInt64(dr[0]), Convert.ToInt64(dr[1]), dr[2].ToString(), Convert.ToInt64(dr[3]), Convert.ToDouble(dr[4]), Convert.ToDateTime(dr[5]), Convert.ToDateTime(dr[6]), Convert.ToInt64(dr[7]), Convert.ToInt64(dr[8]));
             return viaje;
         }
         #endregion
@@ -59,7 +59,5 @@ namespace UberFrba.Registro_Viajes
                     + " WHERE id_viaje =" + Convert.ToString(viaje.Id));
             dr.Close();
         }
-
-
     }
 }
